backend/requirements.txt
# Core
Django>=6.0,<7.0
gunicorn

# Banco de dados
psycopg2-binary
django-environ

# API REST
djangorestframework
django-cors-headers

# Autenticação JWT
djangorestframework-simplejwt

# Segurança e utilitários
django-filter
whitenoise

# Async compatibilidade (se desejar ASGI)
uvicorn[standard]

django-import-export

pandas
openpyxl